+++ Directory.Packages.props
@@ -5,7 +5,7 @@
   </PropertyGroup>
   <ItemGroup>
     <PackageVersion Include="Acontplus.Barcode" Version="1.0.4" />
-    <PackageVersion Include="Acontplus.Core" Version="1.5.2" />
+    <PackageVersion Include="Acontplus.Core" Version="1.5.4" />
     <PackageVersion Include="Acontplus.Persistence.Common" Version="1.1.10" />
     <PackageVersion Include="Acontplus.Utilities" Version="1.3.9" />
     <PackageVersion Include="Asp.Versioning.Mvc" Version="8.1.0" />
