+++ src/Acontplus.Core/GlobalUsings.cs
@@ -1,11 +1,10 @@
-global using System.ComponentModel;
 global using Acontplus.Core.Abstractions.Persistence;
 global using Acontplus.Core.Domain.Common.Events;
-global using Acontplus.Core.Domain.Common.Results;
 global using Acontplus.Core.Domain.Enums;
 global using Acontplus.Core.DTOs.Requests;
 global using Acontplus.Core.DTOs.Responses;
 global using Acontplus.Core.Enums;
+global using System.ComponentModel;
 global using System.Data;
 global using System.Data.Common;
 global using System.Diagnostics;
