+++ src/Acontplus.Services/GlobalUsings.cs
@@ -2,10 +2,18 @@ global using Acontplus.Core.Abstractions.Context;
 global using Acontplus.Core.Domain.Enums;
 global using Acontplus.Core.Domain.Exceptions;
 global using Acontplus.Core.DTOs.Responses;
+global using Acontplus.Services.Configuration;
+global using Acontplus.Services.Extensions.Context;
+global using Acontplus.Services.Extensions.Middleware;
+global using Acontplus.Services.Middleware;
+global using Acontplus.Services.Services.Abstractions;
+global using Acontplus.Services.Services.Implementations;
 global using Microsoft.AspNetCore.Authentication.JwtBearer;
 global using Microsoft.AspNetCore.Authorization;
 global using Microsoft.AspNetCore.Builder;
+global using Microsoft.AspNetCore.Hosting;
 global using Microsoft.AspNetCore.Http;
+global using Microsoft.AspNetCore.Mvc.Filters;
 global using Microsoft.AspNetCore.RateLimiting;
 global using Microsoft.AspNetCore.ResponseCompression;
 global using Microsoft.Extensions.Caching.Memory;
@@ -29,11 +37,3 @@ global using System.Text;
 global using System.Text.Json;
 global using System.Threading.RateLimiting;
 global using System.Threading.Tasks;
-global using Acontplus.Services.Configuration;
-global using Acontplus.Services.Extensions.Context;
-global using Acontplus.Services.Extensions.Middleware;
-global using Acontplus.Services.Middleware;
-global using Acontplus.Services.Services.Abstractions;
-global using Acontplus.Services.Services.Implementations;
-global using Microsoft.AspNetCore.Hosting;
-global using Microsoft.AspNetCore.Mvc.Filters;
