+++ src/Acontplus.Utilities/GlobalUsings.cs
@@ -1,9 +1,5 @@
-global using System;
-global using System.Collections;
-global using System.Collections.Generic;
 global using Acontplus.Core.Constants;
 global using Acontplus.Core.Domain.Common.Results;
-global using Acontplus.Core.Domain.Extensions;
 global using Acontplus.Core.DTOs.Common;
 global using Acontplus.Core.DTOs.Responses;
 global using Acontplus.Core.Enums;
@@ -15,6 +11,9 @@ global using Microsoft.AspNetCore.DataProtection;
 global using Microsoft.AspNetCore.Http;
 global using Microsoft.AspNetCore.Mvc;
 global using Microsoft.Extensions.Configuration;
+global using System;
+global using System.Collections;
+global using System.Collections.Generic;
 global using System.Data;
 global using System.Diagnostics;
 global using System.Globalization;
